feat: api response 수정 반영 및 챌린지 갯수 추가
