+++ Projects/Domain/Sources/Model/Challenge/ChallengeModel.swift
@@ -1,0 +1,8 @@
+//
+//  ChallengeModel.swift
+//  Domain
+//
+//  Created by 유현진 on 8/15/24.
+//
+
+import Foundation
